test: link mcp gateway transport coverage
