+++ scripts/e2e/cron-mcp-cleanup-docker-client.ts
@@ -7,12 +7,12 @@ import path from "node:path";
 import { setTimeout as delay } from "node:timers/promises";
 import { pathToFileURL } from "node:url";
 import { promisify } from "node:util";
+import type { GatewayRpcClient } from "../../test/e2e/qa-lab/runtime/mcp-channels.fixture.ts";
 import { readPositiveIntEnv } from "./lib/env-limits.mjs";
-import type { GatewayRpcClient } from "./mcp-channels-harness.ts";
 
 const execFileAsync = promisify(execFile);
 const PROBE_PID_WAIT_MS = readCronMcpCleanupProbePidWaitMs();
-type McpChannelsHarness = typeof import("./mcp-channels-harness.ts");
+type McpChannelsHarness = typeof import("../../test/e2e/qa-lab/runtime/mcp-channels.fixture.ts");
 let mcpChannelsHarness: McpChannelsHarness | undefined;
 
 type CronJob = { id?: string };
@@ -21,7 +21,7 @@ type AgentRunResult = { runId?: string; status?: string };
 type CronFinishedPayload = { status?: unknown };
 
 async function loadMcpChannelsHarness(): Promise<McpChannelsHarness> {
-  mcpChannelsHarness ??= await import("./mcp-channels-harness.ts");
+  mcpChannelsHarness ??= await import("../../test/e2e/qa-lab/runtime/mcp-channels.fixture.ts");
   return mcpChannelsHarness;
 }
