+++ scripts/e2e/mcp-channels-docker.sh
@@ -66,7 +66,7 @@ docker_e2e_run_with_harness \
     tsx scripts/e2e/mcp-channels-seed.ts >/tmp/mcp-channels-seed.log
     gateway_pid=\"\$(openclaw_e2e_start_gateway \"\$entry\" $PORT /tmp/mcp-channels-gateway.log)\"
     openclaw_e2e_wait_gateway_ready \"\$gateway_pid\" /tmp/mcp-channels-gateway.log 480 $PORT
-    tsx scripts/e2e/mcp-channels-docker-client.ts
+    tsx test/e2e/qa-lab/runtime/mcp-channels-docker-client.ts
   " >"$CLIENT_LOG" 2>&1
 status=${PIPESTATUS[0]}
 set -e
