+++ test/e2e/qa-lab/runtime/mcp-client-temp-state.fixture.ts
@@ -1,4 +1,4 @@
-// Mcp Client Temp State script supports OpenClaw repository automation.
+// MCP client temp-state helpers used by QA-owned MCP E2E fixtures.
 import { mkdirSync, mkdtempSync, rmSync, writeFileSync } from "node:fs";
 import { tmpdir } from "node:os";
 import path from "node:path";
@@ -15,6 +15,12 @@ export type ReconnectableMcpClientHandle = {
   client: { close: () => Promise<unknown> };
   transport: { close: () => Promise<unknown> };
 };
+
+type McpConnectTransport = {
+  close?(): Promise<void> | void;
+};
+
+const MCP_TIMEOUT_CLOSE_GRACE_MS = 5_000;
 
 export function createMcpClientTempState(params: {
   gatewayToken: string;
@@ -33,6 +39,55 @@ export function createMcpClientTempState(params: {
     stateDir,
     tokenFile,
   };
+}
+
+export async function connectMcpWithTimeout<TTransport extends McpConnectTransport>(
+  client: { connect(transport: TTransport): Promise<void> },
+  transport: TTransport,
+  timeoutMs: number,
+): Promise<void> {
+  let timedOut = false;
+  let timeout: NodeJS.Timeout | undefined;
+  const timeoutPromise = new Promise<never>((_, reject) => {
+    timeout = setTimeout(() => {
+      timedOut = true;
+      reject(new Error(`MCP stdio connect timed out after ${timeoutMs}ms`));
+    }, timeoutMs);
+    timeout.unref?.();
+  });
+
+  try {
+    await Promise.race([client.connect(transport), timeoutPromise]);
+  } catch (error) {
+    if (timedOut) {
+      await closeTimedOutTransport(transport);
+    }
+    throw error;
+  } finally {
+    if (timeout) {
+      clearTimeout(timeout);
+    }
+  }
+}
+
+async function closeTimedOutTransport(transport: McpConnectTransport): Promise<void> {
+  if (!transport.close) {
+    return;
+  }
+  let timer: NodeJS.Timeout | undefined;
+  try {
+    await Promise.race([
+      Promise.resolve(transport.close()).catch(() => undefined),
+      new Promise<void>((resolve) => {
+        timer = setTimeout(resolve, MCP_TIMEOUT_CLOSE_GRACE_MS);
+        timer.unref?.();
+      }),
+    ]);
+  } finally {
+    if (timer) {
+      clearTimeout(timer);
+    }
+  }
 }
 
 export async function connectMcpClientWithPairingReconnect<
